requirements.txt
#common
python-dotenv
#openai
openai
#google-geminiai
google-generativeai